+++ src/main/kotlin/dsm/pick2024/domain/notification/event/CreateSubscribeTopicEvent.kt
@@ -1,18 +1,17 @@
-package dsm.pick2024.domain.notification.service
+package dsm.pick2024.domain.notification.event
 
 import dsm.pick2024.domain.event.Topic
 import dsm.pick2024.domain.notification.domain.TopicSubscription
-import dsm.pick2024.domain.notification.port.`in`.CreateSubscribeTopicUseCase
 import dsm.pick2024.domain.notification.port.out.QueryTopicSubscriptionPort
 import dsm.pick2024.domain.notification.port.out.SaveTopicSubscriptionPort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class CreateSubscribeTopicService(
+class CreateSubscribeTopicEvent(
     private val saveTopicSubscriptionPort: SaveTopicSubscriptionPort,
     private val queryTopicSubscriptionPort: QueryTopicSubscriptionPort
-): CreateSubscribeTopicUseCase {
+): CreateSubscribeTopicEventUseCase {
 
     @Transactional
     override fun execute(deviceToken: String) {
