+++ src/main/kotlin/dsm/pick2024/domain/notification/event/CreateSubscribeTopicEventUseCase.kt
@@ -1,0 +1,5 @@
+package dsm.pick2024.domain.notification.event
+
+interface CreateSubscribeTopicEventUseCase {
+    fun execute(deviceToken: String)
+}
