--- src/main/kotlin/dsm/pick2024/domain/notification/port/in/CreateSubscribeTopicUseCase.kt
@@ -1,5 +1,0 @@
-package dsm.pick2024.domain.notification.port.`in`
-
-interface CreateSubscribeTopicUseCase {
-    fun execute(deviceToken: String)
-}
